General work
Finished up a bit of the big room with the bridges, reworked a lot of
geometry to fit inside the old map's clipping (let's see if we can't
make this BSP interchangeable with the base map :D).
Old map clipping is a func_group, just hide it.
Worked on structural brushes a bit.
Finished initial floor below FFA platform.